+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [react] feat: Add useProfiler hook (#2659)
 - [core] feat: Export `makeMain` (#2665)
 - [core] fix: Call `bindClient` when creating new `Hub` to make integrations work automatically (#2665)
+- [gatsby] feat: Add @sentry/gatsby package (#2652)
 
 ## 5.17.0